requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.3
pdfplumber==0.11.4
psycopg2==2.9.10
psycopg2_binary==2.9.10
python_docx==1.1.2
Requests==2.32.3
spacy==3.8.2
streamlit==1.38.0
